feat(subscription api): create stripe custumer and connect to supabase user
When the use (signed in via supabase subscripes, stripe user is created and in supabase table called user-profiles, supabase user id and stripe user id is connected as a foreign key. these gives us a posiibility to identify user who has subscribed and who have not, after which we can allow or deny access on specific routes e.x. dashboard for subscribed users and for not subscribers